pyproject.toml
[tool.poetry]
name = "blog-django"
version = "0.1.0"
description = ""
authors = ["CoderBerryRabbit <CoderChaos@yeah.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.16"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
PyMySQL = "^1.0.2"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.4"
serializers = "^0.2.4"
django-redis = "^5.2.0"
drf-extensions = "^0.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
